backend-django/requirements.txt
Django
psycopg2
django-cors-headers
django-rest-framework
